requirements.txt
matplotlib==3.9.2
numpy==2.1.2
opencv_python==4.10.0.84
Pillow==11.0.0
torch==2.3.1
torch==2.5.0
torchvision==0.20.0
